fieldName {RSocrata}	R Documentation

Convert Socrata human-readable column name to field name

Description

Convert Socrata human-readable column name, as it might appear in the first row of data, to field name as it might appear in the HTTP header; that is, lower case, periods replaced with underscores#'

Usage

fieldName(humanName)

Arguments

humanName
- a Socrata human-readable column name

Value

Socrata field name

Author(s)

Hugh J. Devlin, Ph. D. Hugh.Devlin@cityofchicago.org

Examples

fieldName("Number.of.Stations") # number_of_stations

[Package RSocrata version 1.7.6-3 Index]