isFourByFour {RSocrata}	R Documentation

Checks the validity of the syntax for a potential Socrata dataset Unique Identifier, also known as a 4x4.

Description

Will check the validity of a potential dataset unique identifier supported by Socrata. It will provide an exception if the syntax does not align to Socrata unique identifiers. It only checks for the validity of the syntax, but does not check if it actually exists.

Usage

isFourByFour(fourByFour)

Arguments

fourByFour
- a string; character vector of length one

Value

TRUE if is valid Socrata unique identifier, FALSE otherwise

Author(s)

Tom Schenk Jr tom.schenk@cityofchicago.org

[Package RSocrata version 1.7.4-7 Index]