validateUrl {RSocrata}	R Documentation

Convert, if necessary, URL to valid REST API URL supported by Socrata.

Description

Will convert a human-readable URL to a valid REST API call supported by Socrata. It will accept a valid API URL if provided by users and will also convert a human-readable URL to a valid API URL. Will accept queries with optional API token as a separate argument or will also accept API token in the URL query. Will resolve conflicting API token by deferring to original URL.

Usage

validateUrl(url, app_token)

Arguments

url
- a string; character vector of length one

app_token
- a string; SODA API token used to query the data portal http://dev.socrata.com/consumers/getting-started.html

Value

a - valid Url

Author(s)

Tom Schenk Jr tom.schenk@cityofchicago.org

[Package RSocrata version 1.7.6-3 Index]